+++ CHANGELOG.md
@@ -14,6 +14,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed `wp_set_object_terms` without append bug where terms of a different language or without language were not being cleared.
 - Fixed empty post type on edit.php for posts in `Posts\BulkEdit::bulk_edit_messages`.
 
+### Added
+
+- Added posts list filter for trashed posts without a language.
+
 ## [0.6.4] - 2026-04-16
 
 ### Fixed